Revert "fix(runtime): indexed writes honour a custom array prototype (from #9297) (#9326)"
This reverts commit 1f8ad06fa17ebd3ff353e40407df914e33469989.
