+++ crates/perry-runtime/src/array/mod.rs
@@ -138,9 +138,8 @@ pub use self::immutable::{
     js_array_with, js_arraylike_copy_within,
 };
 pub(crate) use self::indexing::{
-    array_custom_prototype, array_has_own_index, array_iteration_is_exotic,
-    array_iteration_is_exotic_resolved, array_prototype_has_index_flag, array_spec_get,
-    array_spec_has_index, array_spec_set,
+    array_has_own_index, array_iteration_is_exotic, array_iteration_is_exotic_resolved,
+    array_prototype_has_index_flag, array_spec_get, array_spec_has_index, array_spec_set,
 };
 pub use self::indexing::{
     js_array_get_element, js_array_get_element_f64, js_array_get_f64, js_array_get_f64_unchecked,
